Refactoring: Extract TerminationReport to its own file
