+++ core/src/main/scala/stainless/termination/TerminationCallBack.scala
@@ -11,9 +11,9 @@ import utils.CheckFilter
 final class TerminationCallBack(override val context: inox.Context) extends CallBackWithRegistry with CheckFilter {
   private implicit val debugSection = DebugSectionTermination
 
-  override type Report = TerminationComponent.TerminationReport
+  override type Report = TerminationReport
   override val cacheSubDirectory = TerminationComponent.name
-  override def parseReportCache(json: org.json4s.JValue): Report = TerminationComponent.TerminationReport.parse(json)
+  override def parseReportCache(json: org.json4s.JValue): Report = TerminationReport.parse(json)
 
   override def onCycleBegin(): Unit = TerminationComponent.onCycleBegin()
